--- a/.github/workflows/compiler.yaml
+++ b/.github/workflows/compiler.yaml
@@ -217,8 +217,8 @@ jobs:
           # Only push results to the benchmark repository if we ran on a commit
           # to the main branch.
           external-data-json-path: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' && '' || 'data.json' }}
+          auto-push: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' && true || false }}
           github-token: ${{ secrets.BENCHMARK_RESULTS_TOKEN }}
-          auto-push: true
           comment-always: true
           comment-on-alert: true
           alert-threshold: 150%