--- a/compiler/src/compiler/hir.rs
+++ b/compiler/src/compiler/hir.rs
@@ -227,7 +227,7 @@ impl Expression {
         Expression::Symbol("Nothing".to_string())
     }
 }
-#[allow(clippy::derive_hash_xor_eq)]
+#[allow(clippy::derived_hash_with_manual_eq)]
 impl hash::Hash for Expression {
     fn hash<H: hash::Hasher>(&self, state: &mut H) {
         core::mem::discriminant(self).hash(state);
@@ -269,7 +269,7 @@ pub enum Pattern {
         errors: Vec<CompilerError>,
     },
 }
-#[allow(clippy::derive_hash_xor_eq)]
+#[allow(clippy::derived_hash_with_manual_eq)]
 impl hash::Hash for Pattern {
     fn hash<H: hash::Hasher>(&self, state: &mut H) {
         core::mem::discriminant(self).hash(state);
@@ -288,7 +288,7 @@ pub struct Body {
     pub expressions: LinkedHashMap<Id, Expression>,
     pub identifiers: HashMap<Id, String>,
 }
-#[allow(clippy::derive_hash_xor_eq)]
+#[allow(clippy::derived_hash_with_manual_eq)]
 impl hash::Hash for Body {
     fn hash<H: hash::Hasher>(&self, state: &mut H) {
         self.expressions.hash(state);