--- a/compiler/src/compiler/mir.rs
+++ b/compiler/src/compiler/mir.rs
@@ -373,7 +373,7 @@ fn test_multiple_flattening() {
     );
 }
 
-#[allow(clippy::derive_hash_xor_eq)]
+#[allow(clippy::derived_hash_with_manual_eq)]
 impl hash::Hash for Expression {
     fn hash<H: hash::Hasher>(&self, state: &mut H) {
         core::mem::discriminant(self).hash(state);