--- a/compiler/src/vm/heap/object.rs
+++ b/compiler/src/vm/heap/object.rs
@@ -16,6 +16,7 @@ use num_bigint::BigInt;
 use std::{
     collections::{hash_map::DefaultHasher, HashMap},
     hash::{Hash, Hasher},
+    iter,
     ops::Deref,
 };
 
@@ -240,7 +241,7 @@ impl Data {
         }
     }
 
-    pub fn children(&self) -> Vec<Pointer> {
+    pub fn children(&self) -> Box<dyn Iterator<Item = Pointer> + '_> {
         match self {
             Data::Int(_)
             | Data::Text(_)
@@ -248,13 +249,10 @@ impl Data {
             | Data::Builtin(_)
             | Data::HirId(_)
             | Data::SendPort(_)
-            | Data::ReceivePort(_) => vec![],
-            Data::List(List { items }) => items.clone(),
-            Data::Struct(struct_) => struct_
-                .iter()
-                .flat_map(|(a, b)| vec![a, b].into_iter())
-                .collect_vec(),
-            Data::Closure(closure) => closure.captured.clone(),
+            | Data::ReceivePort(_) => Box::new(iter::empty()),
+            Data::List(List { items }) => Box::new(items.iter().copied()),
+            Data::Struct(struct_) => Box::new(struct_.iter().flat_map(|(a, b)| vec![a, b])),
+            Data::Closure(closure) => Box::new(closure.captured.iter().copied()),
         }
     }