--- a/compiler/src/vm/heap/pointer.rs
+++ b/compiler/src/vm/heap/pointer.rs
@@ -16,6 +16,6 @@ impl Pointer {
 }
 impl Display for Pointer {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "0x{:x}", self.0)
+        write!(f, "0x{:X}", self.0)
     }
 }