--- a/old-packages/name_checker/analysis_options.yaml
+++ b/old-packages/name_checker/analysis_options.yaml
@@ -1,49 +1,121 @@
-include: package:extra_pedantic/analysis_options.yaml
+include: package:flutter_lints/flutter.yaml
 
 analyzer:
-  strong-mode:
-    implicit-casts: false
-    implicit-dynamic: false
-  exclude:
-    - build/**
-    - lib/**.g.dart
-    - lib/**.freezed.dart
-    - lib/generated/**
+  language:
+    strict-casts: true
+    strict-inference: true
+    strict-raw-types: true
 
 linter:
   rules:
-    always_put_required_named_parameters_first: false
+    always_declare_return_types: true
     avoid_bool_literals_in_conditional_expressions: true
-    avoid_classes_with_only_static_members: true
+    avoid_catching_errors: true
+    avoid_double_and_int_checks: true
     avoid_field_initializers_in_const_classes: true
-    avoid_function_literals_in_foreach_calls: true
+    avoid_final_parameters: true
+    avoid_implementing_value_types: true
+    avoid_js_rounded_ints: true
+    avoid_multiple_declarations_per_line: true
     avoid_positional_boolean_parameters: true
-    avoid_print: true
+    avoid_returning_this: true
     avoid_setters_without_getters: true
+    avoid_slow_async_io: true
     avoid_types_on_closure_parameters: true
-    camel_case_extensions: true
-    camel_case_types: true
-    cascade_invocations: true
+    avoid_unused_constructor_parameters: true
+    avoid_void_async: true
+    cancel_subscriptions: true
+    cast_nullable_to_non_nullable: true
+    close_sinks: true
+    combinators_ordering: true
     comment_references: true
-    constant_identifier_names: true
-    curly_braces_in_flow_control_structures: false
+    conditional_uri_does_not_exist: true
+    depend_on_referenced_packages: false
+    directives_ordering: true
+    discarded_futures: true
     flutter_style_todos: true
-    literal_only_boolean_expressions: false
-    non_constant_identifier_names: true
+    invariant_booleans: true
+    join_return_with_assignment: true
+    no_adjacent_strings_in_list: true
     omit_local_variable_types: true
-    one_member_abstracts: true
-    overridden_fields: true
-    prefer_const_constructors: false
-    prefer_constructors_over_static_methods: true
-    prefer_function_declarations_over_variables: true
+    only_throw_errors: true
+    prefer_asserts_in_initializer_lists: true
+    prefer_final_in_for_each: true
+    prefer_final_locals: true
+    prefer_foreach: true
+    prefer_if_elements_to_conditional_expressions: true
     prefer_int_literals: true
-    prefer_interpolation_to_compose_strings: true
-    prefer_mixin: true
+    prefer_relative_imports: true
     prefer_single_quotes: true
+    secure_pubspec_urls: true
+    sized_box_shrink_expand: true
     sort_constructors_first: true
-    type_annotate_public_apis: false
-    unnecessary_brace_in_string_interps: true
+    sort_pub_dependencies: true
+    sort_unnamed_constructors_first: true
+    test_types_in_equals: true
+    throw_in_finally: true
+    tighten_type_of_initializing_formals: true
+    unawaited_futures: true
+    unnecessary_await_in_return: false
     unnecessary_lambdas: true
+    unnecessary_null_aware_operator_on_extension_on_nullable: true
+    unnecessary_parenthesis: true
     unnecessary_statements: true
-    unnecessary_this: true
+    unnecessary_to_list_in_spreads: true
+    unreachable_from_main: true
+    unsafe_html: true
+    use_colored_box: true
+    use_decorated_box: true
+    use_enums: true
+    use_key_in_widget_constructors: false
     use_setters_to_change_properties: true
+    use_string_buffers: true
+    use_string_in_part_of_directives: true
+    use_super_parameters: true
+
+dart_code_metrics:
+  rules:
+    - always-remove-listener
+    - avoid-border-all
+    - avoid-cascade-after-if-null
+    - avoid-collection-methods-with-unrelated-types
+    - avoid-duplicate-exports
+    - avoid-global-state
+    - avoid-missing-enum-constant-in-map
+    - avoid-nested-conditional-expressions:
+        acceptable-level: 3
+    - avoid-shrink-wrap-in-lists
+    - avoid-top-level-members-in-tests
+    - avoid-unnecessary-setstate
+    - avoid-unnecessary-type-assertions
+    - avoid-unnecessary-type-casts
+    - avoid-unrelated-type-assertions
+    - avoid-unused-parameters
+    - avoid-use-expanded-as-spacer
+    - avoid-wrapping-in-padding
+    - binary-expression-operand-order
+    - check-for-equals-in-render-object-setters
+    - consistent-update-render-object
+    - double-literal-format
+    - missing-test-assertion
+    - no-boolean-literal-compare
+    - no-equal-then-else
+    - no-object-declaration
+    - prefer-async-await
+    - prefer-const-border-radius # ?
+    - prefer-correct-edge-insets-constructor
+    - prefer-correct-test-file-name
+    - prefer-enums-by-name
+    - prefer-extracting-callbacks:
+        allowed-line-count: 5
+        ignored-named-arguments:
+          - builder
+          - rowBuilder
+    - prefer-first
+    - prefer-immediate-return
+    - prefer-intl-name
+    - prefer-iterable-of
+    - prefer-last
+    - prefer-trailing-comma
+    - provide-correct-intl-args
+    - tag-name