<!-- Please enter the corresponding issue ID: -->

<!-- Add the breaking label (PR: BREAKING) if applicable. -->

<!-- Please summarize your changes: -->
Overall CI results: https://github.com/candy-lang/candy/commit/6107a36b73fe98ef47d0cb29623b6f62efff620e#commitcomment-96072605

---

https://github.com/candy-lang/candy/commit/fcf1f1ec9da4010e86850dabae30e056c67ddf0d (almost 60 % runtime improvement)

Benchmark results for `fibonacci 15` (`hyperfine "cargo run --release --manifest-path=compiler/Cargo.toml -- run packages/examples/fibonacci.candy"`):


```
# Before:
Time (mean ± σ):      1.317 s ±  0.047 s    [User: 1.241 s, System: 0.072 s]
Range (min … max):    1.262 s …  1.395 s    10 runs
# After:
Time (mean ± σ):      1.084 s ±  0.041 s    [User: 1.011 s, System: 0.070 s]
Range (min … max):    1.042 s …  1.166 s    10 runs
```

Criterion results:

```
# Before:
VM Runtime/hello_world  time:   [12.165 ms 12.684 ms 13.266 ms]
VM Runtime/fibonacci/15 time:   [433.30 ms 447.60 ms 463.98 ms]
# After:
VM Runtime/hello_world  time:   [11.718 ms 12.196 ms 12.762 ms]
                        change: [-9.5150% -3.8487% +2.1312%] (p = 0.22 > 0.05)
                        No change in performance detected.
VM Runtime/fibonacci/15 time:   [174.63 ms 180.62 ms 187.04 ms]
                        change: [-61.578% -59.646% -57.677%] (p = 0.00 < 0.05)
                        Performance has improved.
```

---

https://github.com/candy-lang/candy/commit/02a4a9131d5938753d7eb75a9a89120f0aef72ee (minimal change)

Benchmark results for `fibonacci 15` (`hyperfine "cargo run --release --manifest-path=compiler/Cargo.toml -- run packages/examples/fibonacci.candy"`):

```
# Before:
Time (mean ± σ):      1.084 s ±  0.041 s    [User: 1.011 s, System: 0.070 s]
Range (min … max):    1.042 s …  1.166 s    10 runs
# After:
Time (mean ± σ):      1.140 s ±  0.061 s    [User: 1.061 s, System: 0.076 s]
Range (min … max):    1.050 s …  1.248 s    10 runs
```

Criterion results:

```
# Before:
VM Runtime/hello_world  time:   [11.718 ms 12.196 ms 12.762 ms]
VM Runtime/fibonacci/15 time:   [174.63 ms 180.62 ms 187.04 ms]
# After:
VM Runtime/hello_world  time:   [12.123 ms 12.513 ms 12.936 ms]
                        change: [-2.9124% +2.5966% +8.0330%] (p = 0.36 > 0.05)
                        No change in performance detected.
VM Runtime/fibonacci/15 time:   [171.87 ms 177.07 ms 182.49 ms]
                        change: [-6.3907% -1.9674% +2.6408%] (p = 0.41 > 0.05)
                        No change in performance detected.
```


### Checklist
<!-- Please check if your PR fulfills the following requirements: -->

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)