@JonasWanke 2023-01-16
I'm not sure what's going on with GitHub here, but I pushed a commit to address the requested change. It appears in the network tab and is shown on the branch, but is not shown here or in the PR's commit tab. https://github.com/candy-lang/candy/commit/d23ee0e4feadf04dced1cea0e052fc75aa900dbf